Kubernetes manifest files for deploying Traefik v2
These files can be used to deploy Traefik v2 as an ingress controller into Kubernetes. This configuration is based on k3s but should apply to other k8s distributions as well.
This deploys the new Traefik v2 Kubernetes CRD for creating IngressRoute resources to route traffic to services.
This contains configuration and documentation on integrating with LetsEncrypt for the automatic creation of HTTPS certificates.