Add AppArmor profile and run container as non-root user
The Traefik and whoami containers are run as non-root users, using nobody:nogroup in this deployment.
The containers listen on ports > 1024 so that this user can bind to the ports, traffic is still sent to
ports 80 and 443 on the host but the Kubernetes service will route this to ports 8080 and 8443 on the container.

The fsGroup is set to 65534 and the acme.json file gets created (or needs to be updated if already existing) to
the nobody:nogroup so that TLS certificates can continue to be written to it.

An AppArmor profile has been added to the deployment and all capabilities dropped.