Set externalTrafficPolicy to Local
The externalTrafficPolicy setting determines whether or not external traffic
is routed directly to the local node running the pod, or can be routed pods
running on other nodes. Setting this to Local means that the external ip
of the load balancer is the node that it is deployed on, and that the
client source IP address is preserved. If this is set to Cluster (the default)
then the client source IP address is lost.

https://kubernetes.io/docs/tasks/access-application-cluster/create-external-load-balancer/