Revert to apiextensions.k8s.io/v1beta1 API version
The Traefik CRDs in this repository were updated to use apiextensions.k8s.io/v1 but introduced a breaking
change with missing openAPIV3Schema definitions.

This has been reverted to the previous state, with a newly introduced file 001-crd.yaml file containing these.

The CRDs will be updated to the latest apiextensions.k8s.io/v1 versions, including the schema definitions, when
Traefik 2.5 is released containing a predefined set of these.

Additional change to add section in README on commenting out AppArmor annotation for hosts that
do not support this which prevents pods from starting.

Fixes #20