Add init container for cert permissions
Traefik will skip certificate resolvers if the certificates file
dpermissions are more open than 600 (rw only for the user). As per their
Helm chart an init container can be used to set the permissions to 600
on the certificates file so that it is used. Failure to do so will mean
that the default insecure Traefik certificate is presented for requested domains.