Add init container for cert permissions (#25)
Traefik will skip certificate resolvers if the certificates file permissions are more
open than 600 (r/w only for the user).
As per their Helm chart an init container can be used to set the permissions to 600
on the certificates file so that it is used. Failure to do so will mean that the default
insecure Traefik certificate is presented for requested domains and untrusted errors
are shown in the browser and for services attempting to connect.

See https://github.com/traefik/traefik-helm-chart/blob/401c8cdf690cbbc765a935c7279566a13b79a082/traefik/values.yaml#L22 for the
usage of this in the Helm chart.