Add CRDs and RBAC for Traefik 2.5
Traefik 2.5 now supports the new apiextensions.k8s.io/v1 and the CRDs have been updated for this.
This includes the OpenAPI validation schema.

https://doc.traefik.io/traefik/v2.5/reference/dynamic-configuration/kubernetes-crd/

Additional update to Basic Authentication example:

Update example to remove the usage of openssl when base64 encoding the username and password
as this was generating credentials for the Kubernetes secret that did not work and the user was not able
to authenticate using the created username and password.